+++ src/servers/src/otlp/traces.rs
@@ -1,0 +1,32 @@
+// Copyright 2023 Greptime Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use api::v1::RowInsertRequests;
+use opentelemetry_proto::tonic::collector::trace::v1::ExportTraceServiceRequest;
+
+use crate::error::Result;
+
+/// Convert OpenTelemetry traces to GreptimeDB row insert requests
+///
+/// See
+/// <https://github.com/open-telemetry/opentelemetry-proto/blob/main/opentelemetry/proto/trace/v1/trace.proto>
+/// for data structure of OTLP metrics.
+///
+/// Returns `InsertRequests` and total number of rows to ingest
+pub fn to_grpc_insert_requests(
+    _request: ExportTraceServiceRequest,
+) -> Result<(RowInsertRequests, usize)> {
+    // TODO:(fys, bobo)
+    todo!()
+}
