+++ src/servers/src/query_handler.rs
@@ -34,6 +34,9 @@ use common_query::Output;
 use opentelemetry_proto::tonic::collector::metrics::v1::{
     ExportMetricsServiceRequest, ExportMetricsServiceResponse,
 };
+use opentelemetry_proto::tonic::collector::trace::v1::{
+    ExportTraceServiceRequest, ExportTraceServiceResponse,
+};
 use session::context::QueryContextRef;
 
 use crate::error::Result;
@@ -101,4 +104,11 @@ pub trait OpenTelemetryProtocolHandler {
         request: ExportMetricsServiceRequest,
         ctx: QueryContextRef,
     ) -> Result<ExportMetricsServiceResponse>;
+
+    /// Handling opentelemetry traces request
+    async fn traces(
+        &self,
+        request: ExportTraceServiceRequest,
+        ctx: QueryContextRef,
+    ) -> Result<ExportTraceServiceResponse>;
 }
